#18 feat:"core에서 score type integer로 변경"

+++ src/main/java/UIManage/UIManager.java
@@ -121,7 +121,7 @@ public class UIManager {
         return InputReader.getScoreIdx();
     }
 
-    public  double getScoreEdit(int ScoreIdx){
+    public  Integer getScoreEdit(int ScoreIdx){
         view.orderToTypeScore(ScoreIdx);
         return InputReader.getScore();
     }
